backend/requirements.txt
# Railway-compatible stable versions
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.36
psycopg[binary]==3.2.3
alembic==1.14.0
pydantic==2.9.2
httpx==0.28.0
redis[hiredis]==5.2.0
python-dotenv==1.0.1
python-multipart==0.0.17
api-analytics[fastapi]==1.0.15

# Performance optimizations (Railway compatible)
orjson==3.10.11
aiofiles==24.1.0
requests==2.32.3

# Web Search and HTTP
aiohttp==3.10.11

# Authentication (required for auth_standalone.py)
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Speech Processing
openai-whisper==20231117
ffmpeg-python==0.2.0

# TTS Service
gtts==2.6.1

# RAG / Vector Embeddings
sentence-transformers==2.2.2
faiss-cpu>=1.8.0
pypdf==3.17.4
